+++ .github/workflows/pr.yml
@@ -29,7 +29,7 @@ jobs:
       - run: just test --release
 
   audit:
-      name: Audit dependecies
+      name: Audit dependencies
       runs-on: ubuntu-latest
       steps:
         - uses: actions/checkout@v4
