+++ src/checkmark_cli/Cargo.toml
@@ -30,7 +30,3 @@ rayon.workspace = true
 futures.workspace = true
 log.workspace = true
 env_logger.workspace = true
-
-[dev-dependencies]
-assert_cmd.workspace = true
-predicates.workspace = true
