+++ src/checkmark_lint/macro/Cargo.toml
@@ -9,10 +9,7 @@ license.workspace = true
 proc-macro = true
 
 [dependencies]
-checkmark_lint_common.workspace = true
 quote.workspace = true
 syn.workspace = true
 darling.workspace = true
-url-macro.workspace = true
-url.workspace = true
 proc-macro2.workspace = true
